refactor: click cli
