+++ Makefile
@@ -1,5 +1,10 @@
 
-paper:
-	git config core.hooksPath .hooks
+clean:
+	rm -rf data
+	rm -rf logs
+	rm -rf plots
+
+manifolds:
+	python -m src build
