+++ src/svm.py
@@ -1,0 +1,2 @@
+from sklearn.svm import OneClassSVM
+
